moved objects into types.ts
